requirements.backend.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# LangChain family (Python 3.13 compatible)
langchain==1.2.13

# Core RAG building blocks
langchain-text-splitters==1.1.1

# Providers / LLM wrappers
langchain-groq==1.1.2
langchain-openai==1.1.12

# Embeddings
langchain-huggingface==1.2.1
sentence-transformers==5.3.0

# Vector stores
langchain-qdrant==1.1.0
qdrant-client==1.17.1

# Ingestion
pypdf==4.2.0

# Document Management
python-multipart>=0.0.9
python-docx>=1.1.0
openpyxl>=3.1.2
aiofiles>=23.2.1

